requirements-agents.txt
# Congressional Agent Data Ingest & Persona Generation
# Install with: pip install -r requirements-agents.txt

# HTTP & async
requests>=2.28.0
aiohttp>=3.8.0
httpx>=0.24.0

# Retry logic & backoff
tenacity>=8.2.0

# Data models & validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Configuration
python-dotenv>=1.0.0

# Database
neo4j>=5.15.0
duckdb>=0.9.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Web scraping
beautifulsoup4>=4.11.0
lxml>=4.9.0
selenium>=4.10.0  # For JavaScript-heavy scorecards

# LLM interaction (reuse from main requirements.txt)
openai>=1.0.0

# Logging
colorlog>=6.7.0
